Round float position to find place in grid
remove integer representation of position in the grid and just round the floating point position whenever we need to check it.
This makes it so the position in the grid is rounded to the nearest integer rather than always being the floor or ceil of the float value - this helps the realism and intuitiveness of sideways movement when under overhangs.